+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/combobox/ItemComboBoxModel.java
@@ -14,6 +14,7 @@ import java.util.function.Function;
 
 import static is.codion.common.item.Item.item;
 import static java.util.Arrays.asList;
+import static java.util.Collections.emptyList;
 import static java.util.Objects.requireNonNull;
 
 /**
@@ -59,7 +60,7 @@ public final class ItemComboBoxModel<T> extends FilteredComboBoxModel<Item<T>> {
    * @return a new combo box model
    */
   public static <T> ItemComboBoxModel<T> itemComboBoxModel() {
-    return new ItemComboBoxModel<>(null, null);
+    return new ItemComboBoxModel<>(null, emptyList());
   }
 
   /**
@@ -68,7 +69,7 @@ public final class ItemComboBoxModel<T> extends FilteredComboBoxModel<Item<T>> {
    * @return a new combo box model
    */
   public static <T> ItemComboBoxModel<T> itemComboBoxModel(List<Item<T>> items) {
-    return new ItemComboBoxModel<>(null, requireNonNull(items));
+    return new ItemComboBoxModel<>(null, items);
   }
 
   /**
@@ -76,7 +77,7 @@ public final class ItemComboBoxModel<T> extends FilteredComboBoxModel<Item<T>> {
    * @return a new combo box model
    */
   public static <T> ItemComboBoxModel<T> sortedItemComboBoxModel() {
-    return sortedItemComboBoxModel((List<Item<T>>) null);
+    return sortedItemComboBoxModel(emptyList());
   }
 
   /**
@@ -94,7 +95,7 @@ public final class ItemComboBoxModel<T> extends FilteredComboBoxModel<Item<T>> {
    * @return a new combo box model
    */
   public static <T> ItemComboBoxModel<T> sortedItemComboBoxModel(Comparator<Item<T>> comparator) {
-    return new ItemComboBoxModel<>(requireNonNull(comparator), null);
+    return new ItemComboBoxModel<>(requireNonNull(comparator), emptyList());
   }
 
   /**